Took notes from review
