+++ TipCatDotNet.Api/Services/Analitics/IAccountStatsService.cs
@@ -12,5 +12,4 @@ public interface IAccountStatsService
 {
     Task AddOrUpdate(Transaction transaction, CancellationToken cancellationToken = default);
     Task<Result<AccountStatsResponse>> Get(MemberContext memberContext, int accountId, CancellationToken cancellationToken = default);
-    Task<List<FacilityStatsResponse>> GetFacilities(int accountId, DataRates rates, string targetCurrency, CancellationToken cancellationToken = default);
 }
